Fix core dumping on CGC bins
An issue existed where during the core dumping procedure, an ELF
note was created to represent information about the auxiliary
vector. Since CGC has no auxiliary vector an empty note was
created and this broke a number of assumptions. I've fixed the
issue in two places. For one, the note is no longer created.
Additionally, dump_write checks the size before attempting the
write, so passing a size of 0 to dump_write with an invalid
pointer as the buffer argument no longer causes an error.